move email to a new module
