+++ lib/events/shareWith.js
@@ -1,32 +1,21 @@
 "use strict";
 const repo = require("../repo"),
       card = require("../card"),
-      mail = require("mailgun-send"),
-      emailValidator = require("email-validator"),
+      email = require("../email"),
       User = require("../models/user"),
       Repo = require("../models/repo"),
-      htmlMinify = require("html-minifier").minify,
       _ = require("underscore");
-
-mail.config({
-  key: process.env.MAILGUN_TOKEN,
-  sender: "noreply@waltzapp.co",
-  // sender: "noreply@sandbox5cd9ca59d4b840ee8f79240882036873.mailgun.org",
-});
-
-function validateEmails(emails) {
-  return emails.every((email) => emailValidator.validate(email));
-}
 
 // server/SHARE_WITH
 module.exports = function shareWith(action, socket) {
   return new Promise((resolve, reject) => {
-    if (action.emails && validateEmails(action.emails) && action.user && action.repo) {
+    if (action.user && action.repo) {
       return User.findRepo(socket.request.user, action.user, action.repo).then((repo) => {
-        mail.send({
+        console.log(`Sending email invite to ${action.emails.join(', ')} for ${action.user}/${action.repo}`);
+        return email.send({
           subject: `You've been invited look at ${action.user}/${action.repo} on Waltz`,
           recipient: action.emails,
-          body: htmlMinify(`
+          body: `
             <body bgcolor="#eee" style="background:#eee;padding:64px;">
               <img
                 src="http://waltzapp.co/img/logo.svg"
@@ -48,13 +37,11 @@ module.exports = function shareWith(action, socket) {
                 <p>Your pals at Waltz</p>
               </div>
           </body>
-          `, {collapseWhitespace: true}),
+          `,
         });
-        console.log(`Sent email invite to ${action.emails.join(', ')} for ${action.user}/${action.repo}`);
-        resolve();
       });
     } else {
-      reject("The provided emails weren't valid.");
+      reject("Please specify a repo to share.");
     }
   });
 }
